A bit more cleanup and moved the patch methods from bulk service over to libs/. I will rework this in the file pass.
