+++ UI/Web/src/app/series-detail/_components/series-detail/series-detail.component.ts
@@ -36,10 +36,7 @@ import {ToastrService} from 'ngx-toastr';
 import {catchError, debounceTime, forkJoin, Observable, of, ReplaySubject, tap} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {BulkSelectionService} from 'src/app/cards/bulk-selection.service';
-import {
-  EditSeriesModalCloseResult,
-  EditSeriesModalComponent
-} from 'src/app/cards/_modals/edit-series-modal/edit-series-modal.component';
+import {EditSeriesModalComponent} from 'src/app/cards/_modals/edit-series-modal/edit-series-modal.component';
 import {DownloadEvent, DownloadService} from 'src/app/shared/_services/download.service';
 import {UtilityService} from 'src/app/shared/_services/utility.service';
 import {Chapter, LooseLeafOrDefaultNumber, SpecialVolumeNumber} from 'src/app/_models/chapter';
@@ -121,6 +118,8 @@ import {ActionResult} from "../../../_models/actionables/action-result";
 import {CardEntityFactory, ChapterCardEntity, VolumeCardEntity} from "../../../_models/card/card-entity";
 import {CardConfigFactory} from "../../../_services/card-config-factory.service";
 import {EntityCardComponent} from "../../../cards/entity-card/entity-card.component";
+import {ModalResult} from "../../../_models/modal/modal-result";
+import {patchEntitySignal, patchSignalArray} from "../../../../libs/patch";
 
 
 enum TabID {
@@ -489,10 +488,10 @@ class SeriesDetailComponent implements OnInit, AfterContentChecked {
       if (res.effect === 'update') {
         const entityMap = new Map<number, any>(res.entity.map(e => [e.id, e]));
 
-        this.bulkSelectionService.patchSignalArray(this.volumes, entityMap);
-        this.bulkSelectionService.patchSignalArray(this.chapters, entityMap);
-        this.bulkSelectionService.patchSignalArray(this.specials, entityMap);
-        this.bulkSelectionService.patchSignalArray(this.storylineChapters, entityMap);
+        patchSignalArray(this.volumes, entityMap);
+        patchSignalArray(this.chapters, entityMap);
+        patchSignalArray(this.specials, entityMap);
+        patchSignalArray(this.storylineChapters, entityMap);
       }
       this.setContinuePoint();
     });
@@ -903,7 +902,7 @@ class SeriesDetailComponent implements OnInit, AfterContentChecked {
   openEditSeriesModal() {
     const modalRef = this.modalService.open(EditSeriesModalComponent, DefaultModalOptions);
     modalRef.componentInstance.series = this.series();
-    modalRef.closed.subscribe((closeResult: EditSeriesModalCloseResult) => {
+    modalRef.closed.subscribe((closeResult: ModalResult<Series>) => {
       if (closeResult.success) {
         window.scrollTo(0, 0);
       }
@@ -953,13 +952,13 @@ class SeriesDetailComponent implements OnInit, AfterContentChecked {
   }
 
   updateChapter(c: Chapter) {
-    this.bulkSelectionService.patchEntitySignal(this.chapters, c);
-    this.bulkSelectionService.patchEntitySignal(this.specials, c);
-    this.bulkSelectionService.patchEntitySignal(this.storylineChapters, c);
+    patchEntitySignal(this.chapters, c);
+    patchEntitySignal(this.specials, c);
+    patchEntitySignal(this.storylineChapters, c);
   }
 
   updateVolume(c: Volume) {
-    this.bulkSelectionService.patchEntitySignal(this.volumes, c);
+    patchEntitySignal(this.volumes, c);
   }
 
   protected readonly LibraryType = LibraryType;
